requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Data validation
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Database
asyncpg>=0.30.0
sqlalchemy>=2.0.36
pgvector>=0.3.6
greenlet>=3.0.0
alembic>=1.13.0  # Database migrations

# Data processing
pandas>=2.2.3
openpyxl>=3.1.5

# Markdown processing
markdown>=3.7

# PDF generation
weasyprint>=63.0
playwright>=1.49.0  # Fallback PDF generation when WeasyPrint unavailable

# Presentation generation
python-pptx>=1.0.0  # PowerPoint PPTX generation

# AI integrations
openai>=1.55.0
google-genai>=0.3.0
tiktoken>=0.5.0  # Token counting for context management

# Utilities
python-dotenv>=1.0.1
httpx>=0.28.0

# Templating
Jinja2>=3.1.0

# Authentication
PyJWT[crypto]>=2.8.0
bcrypt>=4.0.0
email-validator>=2.0.0
google-auth>=2.35.0

# Error Monitoring
sentry-sdk[fastapi]>=2.19.0

# Rate Limiting
slowapi>=0.1.9
